feat: Material UI theme (#21)
* Add material-ui folder and files

* Add confirm button style

* Add cancel button style

* Replace "none" with 0 opacity color

To override buttons background color

* Use new flex justify variable

The variable have been introduced in sweetalert version 8.11.1

* Set color for disabled buttons

* Upgrade sweetalert2 dependency version

* Add hover for close button

* Add style for select

* Implement hover background for close

Using CSS background property, the circle on hover is slightly
misaligned. To workaround that, an SVG circle is used as background
image. The circle is then moved slightly down to be centered around
the X.

* Add error icon for text box

* Add radio input styling

There are a lot of "magic number" so we need to  review this

* Adjust poisition of swal2-label stylying for radio

Previous code was adding the radio control also for checkboxes

* Add style for range / slider

Added using engine pseudo-elements

* Apply material UI style to checkbox

Apply material UI style to checkbox

* Fix merge issue

* Fix label padding merge issue

Commit 7f14f5c (PR) was created without taking
into consideration commit 76aacc2

* Implement focus for action buttons

* Darken active action button background

Need to be same color as `focus`

* Set font-family for the close button

The serif font-family is not vertically aligned on macOS.
Arial font-family is better vertically aligned

* Merge conflicts with master

* Fix label padding merge issue

Commit 7f14f5c (PR) was created without taking
into consideration commit 76aacc2

* Implement focus for action buttons

* Darken active action button background

Need to be same color as `focus`

* Set font-family for the close button

The serif font-family is not vertically aligned on macOS.
Arial font-family is better vertically aligned

* fix lint errors

* remove .sass-lint.yml